feat(domino): refactor physicsnemo
