+++ examples/domino/test.py
@@ -1,30 +1,18 @@
-# SPDX-FileCopyrightText: Copyright (c) 2023 - 2024 NVIDIA CORPORATION & AFFILIATES.
-# SPDX-FileCopyrightText: All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
+# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
+
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-#
+
 #     http://www.apache.org/licenses/LICENSE-2.0
-#
+
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-This code defines a distributed pipeline for testing the DoMINO model on
-CFD datasets. It includes the instantiating the DoMINO model and datapipe,
-automatically loading the most recent checkpoint, reading the VTP/VTU/STL
-testing files, calculation of parameters required for DoMINO model and
-evaluating the model in parallel using DataParallel across multiple
-GPUs. This is a common recipe that enables training of combined models for surface
-and volume as well either of them separately. The model predictions are loaded in
-the the VTP/VTU files and saved in the specified directory. The eval tab in
-config.yaml can be used to specify the input and output directories.
-"""
+#
+# refs: https://github.com/NVIDIA/physicsnemo/tree/main/examples/cfd/external_aerodynamics/domino
 
 import os
 import re
